Event_Processor/requirements.txt
kafka-python-ng
pydantic==2.5.2
python-dotenv==1.0.0